requirements.txt
streamlit
streamlit_js_eval
requests
folium
python-dotenv
